Fix #90598
